+++ layers/Schema/packages/user-roles/src/ModuleProcessors/FormInputs/FilterInputModuleProcessor.php
@@ -1,0 +1,86 @@
+<?php
+
+declare(strict_types=1);
+
+namespace PoPSchema\UserRoles\ModuleProcessors\FormInputs;
+
+use PoP\ComponentModel\Schema\SchemaDefinition;
+use PoP\ComponentModel\ModuleProcessors\DataloadQueryArgsFilterInputModuleProcessorInterface;
+use PoP\ComponentModel\ModuleProcessors\DataloadQueryArgsSchemaFilterInputModuleProcessorTrait;
+use PoP\ComponentModel\ModuleProcessors\DataloadQueryArgsSchemaFilterInputModuleProcessorInterface;
+use PoP\ComponentModel\ModuleProcessors\AbstractFormInputModuleProcessor;
+use PoPSchema\UserRoles\FilterInputProcessors\FilterInputProcessor;
+
+class FilterInputModuleProcessor extends AbstractFormInputModuleProcessor implements DataloadQueryArgsFilterInputModuleProcessorInterface, DataloadQueryArgsSchemaFilterInputModuleProcessorInterface
+{
+    use DataloadQueryArgsSchemaFilterInputModuleProcessorTrait;
+
+    public const MODULE_FILTERINPUT_USER_ROLES = 'filterinput-user-roles';
+    public const MODULE_FILTERINPUT_EXCLUDE_USER_ROLES = 'filterinput-exclude-user-roles';
+
+    public function getModulesToProcess(): array
+    {
+        return array(
+            [self::class, self::MODULE_FILTERINPUT_USER_ROLES],
+            [self::class, self::MODULE_FILTERINPUT_EXCLUDE_USER_ROLES],
+        );
+    }
+
+    public function getFilterInput(array $module): ?array
+    {
+        $filterInputs = [
+            self::MODULE_FILTERINPUT_USER_ROLES => [FilterInputProcessor::class, FilterInputProcessor::FILTERINPUT_USER_ROLES],
+            self::MODULE_FILTERINPUT_EXCLUDE_USER_ROLES => [FilterInputProcessor::class, FilterInputProcessor::FILTERINPUT_EXCLUDE_USER_ROLES],
+        ];
+        return $filterInputs[$module[1]] ?? null;
+    }
+
+    public function getName(array $module): string
+    {
+        return match ($module[1]) {
+            self::MODULE_FILTERINPUT_USER_ROLES => 'roles',
+            self::MODULE_FILTERINPUT_EXCLUDE_USER_ROLES => 'excludeRoles',
+            default => parent::getName($module),
+        };
+    }
+
+    public function getSchemaFilterInputType(array $module): string
+    {
+        return match ($module[1]) {
+            self::MODULE_FILTERINPUT_USER_ROLES => SchemaDefinition::TYPE_STRING,
+            self::MODULE_FILTERINPUT_EXCLUDE_USER_ROLES => SchemaDefinition::TYPE_STRING,
+            default => $this->getDefaultSchemaFilterInputType(),
+        };
+    }
+
+    public function getSchemaFilterInputIsArrayType(array $module): bool
+    {
+        return match ($module[1]) {
+            self::MODULE_FILTERINPUT_USER_ROLES,
+            self::MODULE_FILTERINPUT_EXCLUDE_USER_ROLES
+                => true,
+            default
+                => false,
+        };
+    }
+
+    public function getSchemaFilterInputIsNonNullableItemsInArrayType(array $module): bool
+    {
+        return match ($module[1]) {
+            self::MODULE_FILTERINPUT_USER_ROLES,
+            self::MODULE_FILTERINPUT_EXCLUDE_USER_ROLES
+                => true,
+            default
+                => false,
+        };
+    }
+
+    public function getSchemaFilterInputDescription(array $module): ?string
+    {
+        $descriptions = [
+            self::MODULE_FILTERINPUT_USER_ROLES => $this->translationAPI->__('Get the users with given roles', 'user-roles'),
+            self::MODULE_FILTERINPUT_EXCLUDE_USER_ROLES => $this->translationAPI->__('Get the users without the given roles', 'user-roles'),
+        ];
+        return $descriptions[$module[1]] ?? null;
+    }
+}